修复属性溢出BUG

+++ gs/src/fire/pb/PChangePointProc.java
@@ -41,11 +41,11 @@ public class PChangePointProc extends mkdb.Procedure{
 			final fire.pb.attr.SRefreshPointType refresh = new fire.pb.attr.SRefreshPointType();			
 			BasicFightProperties bfp = prop.getBfp();
 			
-			refresh.bfp.agi = (short)bfp.getAgi();
-			refresh.bfp.cons = (short)bfp.getCons();
-			refresh.bfp.endu = (short)bfp.getEndu();
-			refresh.bfp.iq = (short)bfp.getIq();
-			refresh.bfp.str = (short)bfp.getStr();			
+			refresh.bfp.agi = bfp.getAgi();
+			refresh.bfp.cons = bfp.getCons();
+			refresh.bfp.endu = bfp.getEndu();
+			refresh.bfp.iq = bfp.getIq();
+			refresh.bfp.str = bfp.getStr();			
 			
 			refresh.bfp.agi_save.putAll(prop.getAddpointfp().getAgi_save());
 			refresh.bfp.cons_save.putAll(prop.getAddpointfp().getCons_save());
